requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy==1.4.27
python-dotenv==0.19.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
google-generativeai==0.3.1
email-validator
python-multipart